+++ Example/Swift/AppDelegate.swift
@@ -7,7 +7,5 @@ import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
     var window: UIWindow?
-
 }